+++ app/assets/javascripts/application.js
@@ -2,6 +2,7 @@
 //= require jquery_ujs
 
 //= require datasets
+//= require normalize_document_counts
 //= require plot_dates
 //= require search
 //= require user
@@ -11,4 +12,4 @@
 //= require jquery.mobile
 
 // Load up the Google Visualization API
-google.load('visualization', '1.0', {'packages':['corechart']});
+google.load('visualization', '1.0', {'packages':['corechart','table']});