+++ app/assets/javascripts/mobileinit.js
@@ -2,6 +2,7 @@
 // Bind all of our various event handlers
 $(document).on("mobileinit", function() {
   bindDatasetsEvents();
+  bindNormalizeDocumentCountsEvents();
   bindPlotDatesEvents();
   bindSearchEvents();
   bindUserEvents();