+++ app/assets/javascripts/normalize_document_counts.js
@@ -1,0 +1,7 @@
+
+function bindNormalizeDocumentCountsEvents() {
+  $(document).on('change', 'select[name="job_params[normalize_doc_counts]"]',
+    function(event, data) {
+      $('#normalize_doc_counts_controls').toggle();
+  });
+}