+++ app/assets/javascripts/plot_dates.js
@@ -2,17 +2,29 @@
 // Graph support for the PlotDates results page
 
 function createPlotDatesGraph() {
+  // Get the elements we need
+  var dataContainer = $.mobile.activePage.find('div.plot_dates_data');
+  var percentContainer = $.mobile.activePage.find('div.plot_dates_percent');
   var graphContainer = $.mobile.activePage.find('div.plot_dates_graph');
-  if (graphContainer.length === 0)
+  var tableContainer = $.mobile.activePage.find('div.plot_dates_table');
+  if (graphContainer.length === 0 || dataContainer.length === 0 ||
+      percentContainer.length === 0 || tableContainer.length === 0)
     return;
 
-  graphContainer.show();
-
+  // Make a DataTable object
   var data = new google.visualization.DataTable();
-  data.addColumn('number', 'Year');
-  data.addColumn('number', 'Documents');
-  data.addRows($.parseJSON(graphContainer.html()));
+  var rows = $.parseJSON(dataContainer.html());
+  var percent = percentContainer.html() == 'true';
 
+  data.addColumn('number', 'Year');
+  if (percent) {
+    data.addColumn('number', 'Fraction of Documents');
+  } else {
+    data.addColumn('number', 'Number of Documents');
+  }
+  data.addRows(rows);
+
+  // Make the line chart object
   var w = $(window).width();
   if (w > 750) {
     w = 750;
@@ -27,13 +39,23 @@ function createPlotDatesGraph() {
 
   var options = { width: w, height: h,
                   legend: { position: 'none' },
-                  hAxis: { format: '####' },
+                  hAxis: { format: '####',
+                           title: 'Year of publication' },
+                  vAxis: { title: 'Number of documents' },
                   pointSize: 3 };
+  if (percent) {
+    options.vAxis.format = '###%';
+    options.vAxis.title = 'Percentage of documents';
+  }
 
   var chart = new google.visualization.LineChart(graphContainer[0]);
   chart.draw(data, options);
 
   graphContainer.trigger('updatelayout');
+
+  // Make a pretty table object
+  var table = new google.visualization.Table(tableContainer[0]);
+  table.draw(data, { page: true, pageSize: 20, sortColumn: 0, width: '20em' });
 }
 
 function bindPlotDatesEvents() {