+++ lib/jobs/analysis/concerns/normalize_document_counts.rb
@@ -25,18 +25,35 @@ module Jobs
       module NormalizeDocumentCounts
         extend ActiveSupport::Concern
 
+        included do
+          # @return [String] if 'on', perform normalization
+          attr_accessor :normalize_doc_counts
+
+          # @return [String] the id of the dataset against which to normalize,
+          #   or blank if the entire corpus
+          attr_accessor :normalize_doc_dataset
+        end
+
         # Normalize a hash containing absolute counts
         #
-        # @param [Dataset] normalization set the dataset to normalize against.
-        #   If +nil+, then normalize against the entire corpus.
+        # @param [User] user the user whose datasets we're querying
         # @param [Symbol] field the field against which to normalize.  This
         #   must obviously match the keys in +counts+.  Can be set to any Solr
         #   field.
         # @param [Hash<String, Integer>] counts the counts of documents,
         #   grouped by +field+ values
         # @param [Hash<String, Float>] the counts of documents, normalized
-        def normalize_document_counts(normalization_set, field, counts)
-          normalize_counts = Solr::DataHeplpers::count_by_field(normalization_set, field)
+        def normalize_document_counts(user, field, counts)
+          return counts unless normalize_doc_counts == 'on'
+
+          if normalize_doc_dataset.blank?
+            @normalization_set = nil
+          else
+            @normalization_set = user.datasets.find(normalize_doc_dataset)
+            fail ArgumentError, 'Normalization set ID is invalid' unless @normalization_set
+          end
+
+          normalize_counts = Solr::DataHelpers::count_by_field(@normalization_set, field)
 
           ret = {}
           counts.each do |k, v|